--- a/src/core/Akka.Tests/Actor/ReceiveActorHandlersTests.cs
+++ b/src/core/Akka.Tests/Actor/ReceiveActorHandlersTests.cs
@@ -12,35 +12,16 @@ using Xunit;
 namespace Akka.Tests.Actor;
 
 public class ReceiveActorHandlersTests
-{   
-    // Tests to do
-    // - Review all tests here as they were AI generated
-    // - Rename the tests to be more accurate.
-    // - Add tests for the AddGenericReceiveHandler method
-    // - Adding for IFoo and then sending a message of Bar : IFoo and it handled
-    // - Decide if tests here should cater for adding receive handlers after "built"
-    // - See if any of the Test_that_signatures_are_equal and Test_that_signatures_differs tests are applicable
-
-
+{
     [Fact]
-    public void Given_a_ReceiveAny_handler_has_been_added_When_adding_any_handler_Then_it_fails()
+    public void Given_ReceiveAnyHandler_Added_When_Adding_Any_Other_Handler_Then_Should_Fail()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddReceiveAnyHandler(_ => { });
 
-        // A ReceiveAny handler has been added, so adding another ReceiveAny handler should fail
-        Assert.Throws<InvalidOperationException>(() => 
+        // A ReceiveAny handler has been added, so adding any other handler should fail
+        Assert.Throws<InvalidOperationException>(() =>
             handlers.AddReceiveAnyHandler(_ => { }));
-    }
-
-    [Fact]
-    public void Given_a_ReceiveAny_handler_has_been_added_When_adding_handler_Then_it_fails()
-    {
-        var handlers = new ReceiveActorHandlers();
-        handlers.AddReceiveAnyHandler(_ => { });
-
-        // A ReceiveAny handler has been added, so adding a handler for object should fail
-        // because ReceiveAny and a receive handler for object are essentially the same
         Assert.Throws<InvalidOperationException>(() =>
             handlers.AddTypedReceiveHandler(typeof(object), null, _ => true));
         Assert.Throws<InvalidOperationException>(() =>
@@ -50,18 +31,18 @@ public class ReceiveActorHandlersTests
     }
 
     [Fact]
-    public void Given_a_TypedReceive_handler_with_predicate_has_been_added_When_adding_any_handler_Then_it_succeeds()
+    public void Given_TypedReceiveHandlerWithPredicate_When_Adding_ReceiveAnyHandler_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddTypedReceiveHandler(typeof(object), _ => true, _ => true);
 
-        // As the object handler has a predicate, adding a ReceiveAny handler should be allowed 
+        // As the object handler has a predicate, adding a ReceiveAny handler should be allowed
         // as the object handler might not handle all objects.
         handlers.AddReceiveAnyHandler(_ => { });
     }
 
     [Fact]
-    public void Given_a_TypedReceive_handler_has_been_added_When_adding_handler_Then_it_fails()
+    public void Given_TypedReceiveHandler_When_Adding_SameTypedReceiveHandler_Then_Should_Fail()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddTypedReceiveHandler(typeof(object), null, _ => true);
@@ -73,7 +54,7 @@ public class ReceiveActorHandlersTests
     }
 
     [Fact]
-    public void Given_a_TypedReceive_handler_with_predicate_has_been_added_When_adding_handler_Then_it_succeeds()
+    public void Given_TypedReceiveHandlerWithPredicate_When_Adding_SameTypedReceiveHandlerWithPredicate_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddTypedReceiveHandler(typeof(object), _ => true, _ => true);
@@ -82,9 +63,35 @@ public class ReceiveActorHandlersTests
         // Adding another handler for the same type combination should be allowed.
         handlers.AddTypedReceiveHandler(typeof(object), null, _ => true);
     }
-    
+
     [Fact]
-    public void Given_a_Generic_handler_with_predicate_has_been_added_When_adding_handler_Then_it_succeeds()
+    public void Given_ObjectTypedReceiveHandlerWithNoPredicate_When_Adding_Any_Other_ReceiveHandler_Then_Should_Fail()
+    {
+        var handlers = new ReceiveActorHandlers();
+        handlers.AddTypedReceiveHandler(typeof(object), null, _ => true);
+
+        // This should throw because the object handler is already added and would catch this before.
+        Assert.Throws<InvalidOperationException>(() =>
+            handlers.AddTypedReceiveHandler(typeof(int), _ => true, _ => true));
+        Assert.Throws<InvalidOperationException>(() =>
+            handlers.AddGenericReceiveHandler<bool>(_ => true, _ => true));
+    }
+
+    // TODO Confirm use case - This is theoretically a breaking change. Conceptually it should not be because Object handler
+    // with no predicate is the same as a ReceiveAny handler.
+    [Fact]
+    public void Given_ObjectTypedReceiveHandlerWithNoPredicate_When_Adding_AnyReceiveHandler_Then_Should_Fail()
+    {
+        var handlers = new ReceiveActorHandlers();
+        handlers.AddTypedReceiveHandler(typeof(object), null, _ => true);
+
+        // This should throw because the object handler is already added and would catch this before.
+        Assert.Throws<InvalidOperationException>(() =>
+            handlers.AddReceiveAnyHandler(_ => { }));
+    }
+
+    [Fact]
+    public void Given_GenericReceiveHandlerWithPredicate_When_Adding_SameGenericReceiveHandlerWithPredicate_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddGenericReceiveHandler<int>(_ => true, _ => true);
@@ -93,22 +100,9 @@ public class ReceiveActorHandlersTests
         // Adding another handler for the same type combination should be allowed.
         handlers.AddGenericReceiveHandler<int>(null, _ => true);
     }
-    
-    
-    [Fact]
-    public void Given_a_Generic_handler_with_predicate_has_been_added_When_adding_handler_Then_it_succeeds1()
-    {
-        var handlers = new ReceiveActorHandlers();
-        handlers.AddGenericReceiveHandler<int>(null, _ => true);
-
-        // The handler has a handler for the type which has no predicate.
-        // Adding another handler for the same type combination should not be allowed.
-        Assert.Throws<InvalidOperationException>(() =>
-            handlers.AddGenericReceiveHandler<int>(_ => true, _ => true));
-    }
 
     [Fact]
-    public void Given_a_TypedReceive_handler_for_different_type_When_adding_handler_Then_it_succeeds()
+    public void Given_TypedReceiveHandler_When_Adding_DifferentTypedReceiveHandler_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddTypedReceiveHandler(typeof(string), _ => true, _ => true);
@@ -117,34 +111,30 @@ public class ReceiveActorHandlersTests
     }
 
     [Fact]
-    public void Given_a_Generic_handler_for_different_type_When_adding_handler_Then_it_succeeds()
+    public void Given_GenericReceiveHandler_When_Adding_DifferentGenericReceiveHandler_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddGenericReceiveHandler<string>(null, _ => true);
 
-        handlers.AddGenericReceiveHandler<int>( _ => true, _ => true);
+        handlers.AddGenericReceiveHandler<int>(_ => true, _ => true);
     }
 
     [Fact]
-    public void Given_a_TypedReceive_handler_with_no_predicate_has_been_added_When_adding_any_handler_Then_it_succeeds()
-    {
-        var handlers = new ReceiveActorHandlers();
-        handlers.AddTypedReceiveHandler(typeof(object), null, _ => true);
-
-        // This should throw because the object handler is already added and would catch this before.
-        Assert.Throws<InvalidOperationException>(() => 
-            handlers.AddTypedReceiveHandler(typeof(int), _ => true, _ => true));
-        Assert.Throws<InvalidOperationException>(() => 
-            handlers.AddGenericReceiveHandler<bool>(_ => true, _ => true));
-    }
-
-    [Fact]
-    public void Given_a_TypedReceive_handler_with_predicate_has_been_added_When_adding_typed_handler_Then_it_succeeds()
+    public void Given_TypedReceiveHandlerWithPredicate_When_Adding_DifferentTypedReceiveHandlerWithPredicate_Then_Should_Succeed()
     {
         var handlers = new ReceiveActorHandlers();
         handlers.AddTypedReceiveHandler(typeof(object), _ => true, _ => true);
-        
-        // This should be allowed  because the object handler is already but it has a predicate that might not match.
+
+        // This should be allowed because the object handler is already but it has a predicate that might not match.
         handlers.AddTypedReceiveHandler(typeof(int), _ => true, _ => true);
     }
+    
+    /*
+     * IFoo
+     * Bar: IFoo
+     *
+     * Receive<IFoo>
+     * Receive<Bar>
+     */
+    
 }