--- a/src/core/Akka.Tests/Actor/ReceiveActorTests.cs
+++ b/src/core/Akka.Tests/Actor/ReceiveActorTests.cs
@@ -1,4 +1,4 @@
-﻿//-----------------------------------------------------------------------
+//-----------------------------------------------------------------------
 // <copyright file="ReceiveActorTests.cs" company="Akka.NET Project">
 //     Copyright (C) 2009-2022 Lightbend Inc. <http://www.lightbend.com>
 //     Copyright (C) 2013-2025 .NET Foundation <https://github.com/akkadotnet/akka.net>
@@ -6,6 +6,7 @@
 //-----------------------------------------------------------------------
 
 using System;
+using System.Threading;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Event;
@@ -138,6 +139,22 @@ namespace Akka.Tests.Actor
             await ExpectMsgAsync(4711);
         }
 
+        [Fact]
+        public async Task Given_an_actor_which_adds_any_handler_twice_should_throw_exception()
+        {
+            // Handling the scenario where the actor adds an any handler twice. This should not be allowed.
+            // Given
+            var system = ActorSystem.Create("test");
+            var actor = system.ActorOf<AnyAddedTwiceActor>("addedtwice");
+            
+            // When
+            using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(3));
+            var terminated =  await actor.WatchAsync(cts.Token);
+
+            // Then
+            Assert.True(terminated);
+        }
+
         private class NoReceiveActor : ReceiveActor
         {
         }
@@ -149,7 +166,6 @@ namespace Akka.Tests.Actor
                 Receive<object>(msg => Sender.Tell(msg, Self));
             }
         }
-
 
         private class PreStartEchoReceiveActor : ReceiveActor
         {
@@ -238,6 +254,21 @@ namespace Akka.Tests.Actor
             }
         }
 
+        private class AnyAddedTwiceActor : ReceiveActor
+        {
+            public AnyAddedTwiceActor()
+            {
+                ReceiveAny(o =>
+                {
+                    Sender.Tell("any:" + o, Self);
+                });
+                ReceiveAny(o =>
+                {
+                    Sender.Tell("not allowed any:" + o, Self);
+                });
+            }
+        }
+        
     }
 }