@mhbuck 2025-02-10
Will expand running the tests locally and narrow it down. 👍