@mhbuck 2025-03-03
Latest Benchmarks

**Before Change**
```
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.5487/22H2/2022Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK 9.0.102
  [Host]     : .NET 8.0.12 (8.0.1224.60305), X64 RyuJIT AVX2
  Job-OKKNQX : .NET 8.0.12 (8.0.1224.60305), X64 RyuJIT AVX2

InvocationCount=1  LaunchCount=10  RunStrategy=Monitoring
UnrollFactor=1  WarmupCount=10

| Method                                | Mean     | Error   | StdDev  | Req/sec      |
|-------------------------------------- |---------:|--------:|--------:|-------------:|
| Actor_ping_pong_single_pair_in_memory | 227.9 ns | 1.43 ns | 4.23 ns | 4 387 834,09 |

Ping Pong
OSVersion:              Microsoft Windows NT 10.0.19045.0
ProcessorCount:         8
ClockSpeed:             0 MHZ
Actor Count:            16
Messages sent/received: 30000000  (3e7)
Is Server GC:           True
Thread count:           26

ActorBase    first start time:  7.23 ms
ReceiveActor first start time: 18.99 ms

            ActorBase                          ReceiveActor
Throughput, Msgs/sec, Start [ms], Total [ms],  Msgs/sec, Start [ms], Total [ms]
         1, 24057000,     111.05,    1358.32,  24469000,     115.64,    1341.86
         5, 39062000,      16.47,     784.60,  39893000,       4.22,     756.33
        10, 43604000,       3.39,     692.36,  45454000,       4.28,     664.38
        15, 57034000,       2.77,     528.90,  51993000,       2.32,     579.61
        20, 49586000,      23.89,     629.86,  52356000,       3.45,     576.73
        30, 58365000,       1.58,     515.81,  46801000,       1.37,     642.70
        40, 62111000,       1.34,     485.30,  51546000,       1.45,     584.12
        50, 52910000,       1.47,     568.76,  49180000,       2.03,     612.23
        60, 61099000,       1.56,     493.42,  54744000,       1.66,     550.64
        70, 59055000,       1.42,     509.71,  59055000,       1.39,     509.48
        80, 49586000,       1.56,     606.87,  49423000,       1.37,     608.59
        90, 64102000,       1.43,     470.09,  57581000,       1.39,     522.40
       100, 62893000,       1.39,     479.06,  56285000,       1.70,     535.58
       200, 57803000,       1.36,     520.62,  55452000,       1.68,     543.05
       300, 53475000,       1.76,     563.48,  58593000,       1.61,     513.69
       400, 57915000,       1.84,     520.35,  59880000,       1.64,     503.34
       500, 55658000,       1.33,     540.33,  60606000,       2.40,     498.05
       600, 64239000,      18.50,     485.94,  60120000,       1.37,     501.04
       700, 58252000,       1.55,     517.49,  56925000,       1.57,     528.88
       800, 60606000,       1.35,     497.16,  52910000,       1.54,     568.78
       900, 59642000,       2.12,     505.87,  57471000,       1.68,     523.71
```
**PR Benchmark**
```
BenchmarkDotNet v0.13.12, Windows 10 (10.0.19045.5487/22H2/2022Update)
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET SDK 9.0.102
  [Host]     : .NET 8.0.12 (8.0.1224.60305), X64 RyuJIT AVX2
  Job-QTOHXP : .NET 8.0.12 (8.0.1224.60305), X64 RyuJIT AVX2

InvocationCount=1  LaunchCount=10  RunStrategy=Monitoring
UnrollFactor=1  WarmupCount=10

| Method                                | Mean     | Error   | StdDev  | Req/sec      |
|-------------------------------------- |---------:|--------:|--------:|-------------:|
| Actor_ping_pong_single_pair_in_memory | 240.2 ns | 1.22 ns | 3.59 ns | 4 162 677,83 |



OSVersion:              Microsoft Windows NT 10.0.19045.0
ProcessorCount:         8
ClockSpeed:             0 MHZ
Actor Count:            16
Messages sent/received: 30000000  (3e7)
Is Server GC:           True
Thread count:           26

ActorBase    first start time: 15.90 ms
ReceiveActor first start time: 17.30 ms

            ActorBase                          ReceiveActor
Throughput, Msgs/sec, Start [ms], Total [ms],  Msgs/sec, Start [ms], Total [ms]
         1, 20775000,      35.37,    1480.08,  24732000,     152.29,    1366.12
         5, 44182000,     139.31,     818.86,  36496000,       2.95,     825.61
        10, 49180000,       7.30,     617.84,  40214000,       5.45,     751.57
        15, 50933000,       2.34,     591.53,  44709000,      12.84,     684.64
        20, 56285000,      97.15,     630.63,  43165000,       4.28,     699.96
        30, 56925000,       1.31,     528.57,  45180000,       2.19,     666.69
        40, 56818000,       2.05,     530.58,  47318000,       1.43,     636.42
        50, 66964000,       1.68,     449.80,  53763000,       1.43,     560.37
        60, 54744000,       1.39,     549.80,  55658000,       1.71,     541.37
        70, 68493000,       1.31,     439.34,  45317000,       5.20,     667.51
        80, 65645000,       1.31,     458.73,  54446000,       1.33,     552.70
        90, 68337000,       3.65,     442.69,  56390000,       1.37,     533.67
       100, 66815000,       1.26,     450.27,  52264000,       1.47,     576.35
       200, 68337000,       1.40,     440.62,  57581000,       1.60,     522.96
       300, 70422000,       1.30,     427.42,  56710000,       1.80,     531.06
       400, 58139000,       1.69,     517.74,  57142000,       1.76,     527.69
       500, 69767000,       1.25,     431.72,  55658000,       1.29,     540.82
       600, 70093000,       2.51,     430.69,  56390000,      22.96,     555.07
       700, 61349000,       1.35,     490.61,  55452000,       1.30,     543.24
       800, 69124000,       1.25,     435.32,  52356000,       1.33,     575.12
       900, 70257000,       1.22,     428.49,  57471000,       1.28,     523.76
```