@Aaronontheweb 2025-03-19
PR benchmark numbers actually look like an improvement, which doesn't surprise me - the stuff we created via `Reflection.Emit` doesn't benefit from PGO and other JIT improvements.